Move project-wide models to shared_models
