+++ botx/bot/api/commands/base.py
@@ -3,14 +3,11 @@ from uuid import UUID
 
 from pydantic import validator
 
-from botx.api_base_models import APIBaseModel
+from botx.shared_models.api_base import APIBaseModel
 from botx.bot.api.constants import BOT_API_VERSION
-from botx.bot.api.enums import (
-    BotAPIChatTypes,
-    BotAPIClientPlatforms,
-    BotAPICommandTypes,
-)
+from botx.bot.api.enums import BotAPIClientPlatforms, BotAPICommandTypes
 from botx.bot.api.exceptions import UnsupportedBotAPIVersionError
+from botx.shared_models.enums import APIChatTypes
 
 try:
     from typing import Literal
@@ -49,7 +46,7 @@ class BotAPIBaseSender(APIBaseModel):
 class BotAPIServerEventSender(BotAPIBaseSender):
     is_admin: Literal[None]
     is_creator: Literal[None]
-    chat_type: Optional[BotAPIChatTypes]
+    chat_type: Optional[APIChatTypes]
     group_chat_id: Optional[UUID]
     user_huid: Optional[UUID]
 
@@ -57,7 +54,7 @@ class BotAPIServerEventSender(BotAPIBaseSender):
 class BotAPIChatEventSender(BotAPIBaseSender):
     is_admin: Literal[None]
     is_creator: Literal[None]
-    chat_type: BotAPIChatTypes
+    chat_type: APIChatTypes
     group_chat_id: UUID
     user_huid: Optional[UUID]
 
@@ -65,7 +62,7 @@ class BotAPIChatEventSender(BotAPIBaseSender):
 class BotAPIUserEventSender(BotAPIBaseSender):
     is_admin: bool
     is_creator: bool
-    chat_type: BotAPIChatTypes
+    chat_type: APIChatTypes
     group_chat_id: UUID
     user_huid: UUID
