+++ botx/client/bots_api/get_token.py
@@ -2,7 +2,7 @@ from typing import NoReturn
 
 import httpx
 
-from botx.api_base_models import APIBaseModel
+from botx.shared_models.api_base import APIBaseModel
 from botx.client.botx_method import BotXMethod
 from botx.client.exceptions import InvalidBotAccountError
