+++ botx/client/botx_api_client.py
@@ -4,7 +4,6 @@ from uuid import UUID
 import httpx
 
 from botx.bot.bot_accounts_storage import BotAccountsStorage
-from botx.bot.models.commands.enums import ChatTypes
 from botx.client.chats_api.create_chat import BotXAPICreateChatPayload, CreateChatMethod
 from botx.client.chats_api.list_chats import ChatListItem, ListChatsMethod
 from botx.client.get_token import get_token
@@ -12,6 +11,7 @@ from botx.client.notifications_api.direct_notification import (
     BotXAPIDirectNotificationPayload,
     DirectNotificationMethod,
 )
+from botx.shared_models.enums import ChatTypes
 
 
 class BotXAPIClient:
