+++ botx/client/chats_api/create_chat.py
@@ -1,10 +1,13 @@
 from typing import List, Optional
 from uuid import UUID
 
-from botx.api_base_models import APIBaseModel
-from botx.bot.models.commands.enums import ChatTypes
+from botx.shared_models.api_base import APIBaseModel
 from botx.client.authorized_botx_method import AuthorizedBotXMethod
-from botx.client.enums import BotXAPIChatTypes, convert_chat_type_from_domain
+from botx.shared_models.enums import (
+    APIChatTypes,
+    ChatTypes,
+    convert_chat_type_from_domain,
+)
 
 try:
     from typing import Literal
@@ -15,7 +18,7 @@ except ImportError:
 class BotXAPICreateChatPayload(APIBaseModel):
     name: str
     description: Optional[str]
-    chat_type: BotXAPIChatTypes
+    chat_type: APIChatTypes
     members: List[UUID]
     shared_history: bool
